refactor: 디렉토리 구조 변경 및 즐겨찾기 모아보기 관련 분기 추가
